ai_server/requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
hanspell==1.1
keybert==0.8.3
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
python-docx==0.8.11
openai==1.3.0
python-dotenv==1.0.0
huggingface-hub==0.17.3
kss==4.5.4
pydantic==2.4.2